new folder structure modified train, test scripts
