+++ examples/rl_frameworks/saved_agents/dqn_64_64.json
@@ -1,0 +1,1 @@
+{"agent": "dqn", "states": {"type": "float", "shape": [17]}, "actions": {"type": "int", "shape": [], "num_values": 7}, "memory": 200000, "batch_size": 25, "max_episode_timesteps": 10000, "network": [{"type": "dense", "size": 64, "activation": "relu"}, {"type": "dense", "size": 64, "activation": "relu"}, {"type": "linear", "size": 7}], "update_frequency": 1, "start_updating": 10000, "learning_rate": 0.0001, "huber_loss": 0.0, "horizon": 0, "discount": 0.99, "estimate_terminal": false, "target_sync_frequency": 1000, "target_update_weight": 1.0, "preprocessing": null, "exploration": {"type": "decaying", "decay": "polynomial", "decay_steps": 50000, "unit": "timesteps", "initial_value": 1.0, "decay_rate": 0.05, "final_value": 0.05, "power": 3.0}, "variable_noise": 0.0, "l2_regularization": 0.0, "entropy_regularization": 0.0, "name": "agent", "device": null, "parallel_interactions": 1, "seed": null, "execution": null, "saver": null, "summarizer": null, "recorder": null, "config": null, "internals": {}, "initial_internals": {}}
