--- examples/tensorforce_for_gem/__init__.py
@@ -1,1 +1,0 @@
-from .tensorforce_utils import set_plot_params, SqdCurrentMonitor, EpsilonWrapper
